sync-eth: add db commands stats, compact, put, get, delete

+++ common/fdlimit/fdlimit_bsd.go
@@ -14,15 +14,18 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with the go-ethereum library. If not, see <http://www.gnu.org/licenses/>.
 
-// +build linux netbsd openbsd solaris
+// +build freebsd dragonfly
 
 package fdlimit
 
 import "syscall"
 
+// This file is largely identical to fdlimit_unix.go,
+// but Rlimit fields have type int64 on *BSD so it needs
+// an extra conversion.
+
 // Raise tries to maximize the file descriptor allowance of this process
 // to the maximum hard-limit allowed by the OS.
-// Returns the size it was set to (may differ from the desired 'max')
 func Raise(max uint64) (uint64, error) {
 	// Get the current limit
 	var limit syscall.Rlimit
@@ -31,17 +34,16 @@ func Raise(max uint64) (uint64, error) {
 	}
 	// Try to update the limit to the max allowance
 	limit.Cur = limit.Max
-	if limit.Cur > max {
-		limit.Cur = max
+	if limit.Cur > int64(max) {
+		limit.Cur = int64(max)
 	}
 	if err := syscall.Setrlimit(syscall.RLIMIT_NOFILE, &limit); err != nil {
 		return 0, err
 	}
-	// MacOS can silently apply further caps, so retrieve the actually set limit
 	if err := syscall.Getrlimit(syscall.RLIMIT_NOFILE, &limit); err != nil {
 		return 0, err
 	}
-	return limit.Cur, nil
+	return uint64(limit.Cur), nil
 }
 
 // Current retrieves the number of file descriptors allowed to be opened by this
